+++ src/test/pegtl/contrib_limit_depth.cpp
@@ -2,6 +2,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 
+#include <tao/pegtl/contrib/input_with_depth.hpp>
 #include <tao/pegtl/contrib/limit_depth.hpp>
 
 #include "test.hpp"
@@ -28,20 +29,22 @@ namespace tao::pegtl
 
    void unit_test()
    {
-      memory_input<> i1( "aaa", __FUNCTION__ );
+      using memory_input_with_depth = input_with_depth< memory_input<> >;
+
+      memory_input_with_depth i1( "aaa", __FUNCTION__ );
       const auto r1 = pegtl::parse< test_grammar >( i1 );
       TAO_PEGTL_TEST_ASSERT( r1 );
 
-      memory_input<> i2( "aaaaaaaaaaa", __FUNCTION__ );
+      memory_input_with_depth i2( "aaaaaaaaaaa", __FUNCTION__ );
       const auto r2 = pegtl::parse< test_grammar >( i2 );
       TAO_PEGTL_TEST_ASSERT( r2 );
 
-      memory_input<> i3( "aaa", __FUNCTION__ );
+      memory_input_with_depth i3( "aaa", __FUNCTION__ );
       const auto r3 = pegtl::parse< test_grammar, test_action >( i3 );
       TAO_PEGTL_TEST_ASSERT( r3 );
 
 #if defined( __cpp_exceptions )
-      memory_input<> i4( "aaaaaaaaaaa", __FUNCTION__ );
+      memory_input_with_depth i4( "aaaaaaaaaaa", __FUNCTION__ );
       TAO_PEGTL_TEST_THROWS( pegtl::parse< test_grammar, test_action >( i4 ) );
 #endif
    }